break up implementation into modules
